core: Decouple networks & bonds names from list
In SetupNetworksHelper the modified networks were kept in a map,
but in order to send just the networks that changed, we need
to decouple the set of networks that are still attached to NIC
from the list of networks that were actually modified.

A similar change is done with the bonds map, but the field needs to
stay as a map in order to be able to process bonds correctly (in next
patches).

Change-Id: Ie324b1acd9022bc3d5f5f0a200040715954132bb
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>