core:Add connection to DB after connectVds(#815083)
https://bugzilla.redhat.com/show_bug.cgi?id=815083

When user adds a faulty storage connection it will cause the command to
fail, but will leave the connection in the data base.

The proposed fix is to add the connection to the DB, only after connect
has succeeded at the execute section.

Also some unused code is being removed, when we get Vds Id which is
empty, rest and GUI should not send an empty vds Id.

Change-Id: I6e1387295f4c24f4222b57924ec2f4380411c5f4
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>