userportal: When can not take VM from pool, UP throws an exception
When tring to take a VM from a pool, but it did not succeeded
(e.g. because there are no available VMs in the pool), than the
UP throws an UnsupportedOperationException, the model gets in
incorrect state and the UP is not functional any more.

The problem was, that the PoolItemBehaviour.TakeVm() was
not checking the result of the AttachUserToVmFromPoolAndRun
callback and tried to work with the taken VM immediately.
Because the taking of the VM failed, the VM was null and the
UserPoralItemModel.OnEntityChanged() did throw an
UnsupportedOperationException.

Solved by adding a check if the operation succeeded to the
PoolItemBehaviour.TakeVm() and continue only if yes.

Change-Id: If110f01cf2abcd8866ac970b6ac4dabf2df7c4cc
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>