userportal: spice client properties
Added the "Enable WAN Options" checkbox to the
edit console popup. It is visible, when all of
the following conditions are met:
- the selected VM is a windows VM
- the spice is available
- the spice radio button is selected
- the spice guest agent is installed on the VM

If the above conditions are met, and the checkbox is
checked, the "WANDisableEffects" and "WANColorDepth"
from the vdc_options table are set to the spice plugin
as the "DisableEffects" and "ColorDepth" properties.

In other case, nothing is set to the "DisableEffects"
and "ColorDepth" properties.

Change-Id: Ic8aa78cea7a05c9edf3fae084b2fdbb4b6943598
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>