packaging: Remove components.war
These web application existed only to hold copies of the scripts
contained in the vdsm-bootstrap packages. Those files are copied
by one of the postinstallation scripts of the engine packages.

This patch removes that application and replaces it with a servlet
in the root web application that reads the scripts from their
original locations.

This also opens interesting possibilities, as that servlet could
be smarter (in the future) and return different versions of the
bootstrap scripts to different versions of hypervisors, or lookup
the files in diferent places.

Change-Id: I07601d8ac1e2c594c1e4098907eedba632b583e2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>