webadmin: Missing gateway field when creating bond (#815415)
https://bugzilla.redhat.com/815415

Bond button for "rhevm" network with vlan tagging does not provide
default gateway field. Problem is originated due to wrong resolution
of the management network when it is on top of a VLAN. This causes
the gateway field not to appear.

With this patch, the gateway field appears and the default values
for the Static IP protocol are the same as were for the management
network interface, those fields are editable.

Change-Id: I42547ef630feda19f520b0eda99bf58d31db4a24
Signed-off-by: Moti Asayag <masayag@redhat.com>