userportal: NPE when there is no Blank template
The problem was, that the it was expected, that
there has to be always a blank template. If there
was not, an empty instance of VmTemplate has been
added to the list of templates which could cause
NPEs.

Solved by adding the blank template to the beginning
of the list only if it is defined (in AsyncDataProvider
and UserPortalNewVmModelBehavior).

The corresponding method from DataProvider has been deleted
as it was never used (and is synchronous which has
no meaning in GWT world)

Change-Id: Ib7ae99cfb3bd663b6fdce373501bed98152b8b6d
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>