core: fix restore from snapshot
A previous commit (cd3dedff2755753a77da01ecf86332c4576a9acc) removed the code that removes all devices, and added
code that only removes disks that are allowed to be in a snapshot.
This caused the devices to be duplicated on every snapshot restore.
This patch fixes this, and removes all devices besides disks that are
not on snapshots.

Change-Id: I7866d0c55153b65023e18075bb57bd7fdfbe78f4
Signed-off-by: Oved Ourfali <oourfali@redhat.com>