packaging: Fix ownership of lock directory
The directory /var/lock/ovirt-engine should be owned by the ovirt
user and group, and it is created that way during the installation
of the package, but changed later during systemctl startup as
specified in the /etc/tmpfiles.d/ovirt-engine.conf file. This
patch changes that file so that the directory is owned by ovirt
and not jboss-as.

Change-Id: I6841dfd987ca7c02c699fcce283a2fcc7e957b88
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>