core: Inconsistent and confusing storage domain statuses (#829275)
https://bugzilla.redhat.com/829275

The following patch will solve a problem of switch storage domain status from Maintaice to Inactive.
The Maintaince status is not known for host, it is only for engine use, at that case a storage domain
status at host side is InActive.
At such case engine will not update status of domain

Change-Id: I94b0d505cd7deed0c19317f2d915c0629e147d99
Signed-off-by: Michael Kublin <mkublin@redhat.com>