webadmin: Import of an existing template show two dialogs
When importing a template that already exists on the data domain, GUI
first shows this: "Import process has begun for Template(s): . You can
check import status in the 'Events' tab of the specific destination
storage domain, or in the main 'Events' tab"

Then this:
"Import Template failed - Template Id already exist in the system.

Only the 2nd message should be shown to the user.

This bug is closly related to the already fixed:
http://gerrit.ovirt.org/#/c/1064
and it is fixed in the same way.
In the future it should be generalized to a common parent class.

Change-Id: I36c86320893abeb097e76db7aa444921c203e100
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>