engine: Allow to iterate over remaining ldap servers on NullPointerException.
When engine-manage-domain is run to add ldap server, current code finds all
ldap servers in the domain and then finds guid on each. In some cases it
gets NullPointerException because the "uniqueId" is not found. In those cases
we currently go to the default "Exception" and break out of ldap server loop.
In my view, we should log/print the message and continue to serve the loop.

This patch is based on http://gerrit.ovirt.org/#/c/4439/2. Since now we have
support for ITDS, the original patch has been updated.

Change-Id: Iaf89ef5d423cde2e420d6196b38506a23416abb8
Signed-off-by: Sharad Mishra <snmishra@linux.vnet.ibm.com>