core: UnPowerMock VdsUpdateRunTimeInfoTest
Remover PowerMock depedency from VdsUpdateRunTimeInfoTest.

In order to acheive this improvement, some refactoring was required in
EjbUtils and VdsUpdateRunTimeInfo intself.

EjbUtils is mocked by setting the inner strategy object and returning it
back to the original obejct after the test is done. If this method is
deemed acceptable, perhaps it would be a good idea to extrat it to its
own @Rule.

Change-Id: I5519a02c197b0ef09a9cabd37e6fa90dbecc0897
Signed-off-by: Allon Mureinik <amureini@redhat.com>