core: adding login permissions
This patch adds the login action group, relevant to LoginUserCommand and
LoginAdminCommand.

In this patch:
1. Adding the action group.
2. Adding a "bottom" object type + object (used for the permissions
check, as every object is the parent of bottom). If there is an object
with permissions for a specific user, and the role contains the login
action group, then the user can login to the system. Otherwise, the user
won't be able to login.
3. Upgrade: The login action group will be added to every predefined role (except for the TemplateUser and the
QuotaUser roles).
4. Common Stored procedures - added a new function that adds an action
group to all the roles, except ones we would like to filter.
5. UI changes - add the new action group under system configuration in
the configuration menu. Also, mark the login action group as a user
action group in the UI.

Change-Id: Ib86ec99739ff73fa500a9de6c056b324c353929c
Signed-off-by: Oved Ourfali <oourfali@redhat.com>