engine: Avoid PM/storage alerts for GlusterOnly clusters
https://bugzilla.redhat.com/829842
https://bugzilla.redhat.com/830820

Bugs:
 - Activation of server triggers "Failed to connect host to storage servers" event
 - Power Management related events/alerts should be removed

This was happening because various class level boolean variables indicating errors related to
storage pool and fencing are set to false by default, and trigger corresponding audit logs/
events when getAuditLogTypeValue() is called. This should not happen if the host belongs to
a cluster that does not support virtualization e.g. gluster cluster.

Also modified VdsCommand to generate the power management alert only when virt service
is enabled on the cluster.

Change-Id: I22d93df0cf6b2594d7eabf972c038baf893eccdf
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>