restapi: Handle PermGen leak for Maven 3
This package uses PowerMock (extensively) in it's tests, which
unfortunately leak PermGen memory. This went unnoticed in Maven 2, but
in Maven 3, which has larger memory requirements, causes the build to
consistantly fail on NicMapperTest.

In order to avoid this problem the tests were forked in a similar way as
in the bll projcet.

Change-Id: I1f6e07281ed4b1653aae7e4c579ec6d587c6987f
Signed-off-by: Allon Mureinik <amureini@redhat.com>