core: Handle correct NIC ID from OVF (#827061)
https://bugzilla.redhat.com/827061

Someone forgot to handle NIC ID correctly, although they changed the
OVF format.
ID was added to NIC in the new OVF format, so need to set new ID only
if the OVF didn't contain ID for the nic.
Overriding the ID always would cause the VM Device for the NIC to be
added but not updated with actual data that was saved in the OVF.

Change-Id: Id842c8b4546e80676f8e8e023e0e66ff9cec5814
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>