core: adding maven step for upgrade script name validation
This patch adds a pom.xml under dbscripts that uses the maven exec
plugin to execute a bash script that checks for upgrade scripts
duplications.
manager pom.xml was modified to include the new dbscripts module and
compile it as first step to throw the error as soon as possible before
other steps are issued.

Change-Id: I26146122acc0b6dfaaea73b232fa3c42a7baff18
Signed-off-by: Eli Mesika <emesika@redhat.com>