packaging: Add key-alias to SSL connector
The "key-alias" attribute for the SSL connector has a default
value of "jboss", but the application server used to take the
first key pair from the keystore when it was missing. Starting
with version 7.1.2 it will always use it, and thus fail with our
keystore which only contains the "engine" alias. This patch
changes the setup tool so that it will always add the attribute.

Change-Id: I2921ad93b2b10de935ae98ef8bc5815e5aa4ecee
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>