core: Change removed bonds to strings
No need to hold the whole interface, hold just the names of the bonds
that got removed.
A set was used since each bond name appears on more than one past slave
so this collection type is better for the job. In the future it will be
used to check if a bond name is contained within..

Change-Id: Id2684e962d2e99dcc8e1c609fdeef079fc290331
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>