core: OVF writer should not depend on LANG setting
The XML writer used to write OVF doesn't specify the character
encoding when creating files. As a result the system selects the
encoding according to the setting of the LANG variable. This will
usually result in selecting UTF-8, but if it happens to be ASCII
the StAX writer will fail and complain that the header of the XML
document doesn't match the encoding of the string. This patch
changes the XML writer so that it always uses the UTF-8 encoding,
regardless of the LANG setting.

Change-Id: I2d3f1693cb0f1c2490360350936a4bff09549e72
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>