userportal: Console popup dialog fixes
Previously, the ConsolePopup was implemented independently
of the standard popup architecture, because the popup
did not have a model class. This caused several problems,
such as the enter/escape binding, title of the popup and
most recently after the console popup has been closed, no
other popups could be opened.

This patch solved this by refactoring the console popup
to fit into the standard popup architecture with it's
own model (UserPortalConsolePopupModel) and commands
for save and close.

This patch also includes the adding the method
beforeCommandExecuted to AbstractModelBoundPopupPresenterWidget
which is executed before every command on that popup
is executed.

Change-Id: I7b5a91a67c703bee7536b3ac5158e1eb35275ebf
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>