core: Cleanup LUN pathes parsing
This patch unifies the code that parses LUN paths from the two places it
was separated in GetDeviceListVDSCommand.
A unit test for the path parsing was also added to make sure no
functionallity was harmed.

In addition, some general cleanups (replacing FQCNs with imports,
extracting constants, etc.) was done in that area.

Change-Id: Ife28fb2d0326fbaaf4a8562d2d709351c75aee81
Signed-off-by: Allon Mureinik <amureini@redhat.com>