engine: Change SetupNetworks diff recognition
SetupNetworks would recognize that NIC definitions changed by comparing
the new NIC from parameters, to the old NIC from the DB. If the NIC is
different then the network/bond is a candidate to be sent to the Host
as changed.

This approach is problematic if a parameter that is not being sent to
the Host is changed, and in turn causes a "false positive" recognition.

Changed the difference recognition to take into account only parameters
that are later on passed to the host:
  Network name, Boot protocol, Specific parameters for Static:
  IP, Gateway, Netmask.

The bond recognition was changed to take into account only when the
bond itself had been modified (either added or changed).
This will allow changing networks & bonds individually, without needing
to change the underlying devices if they didn't change.

Change-Id: I1908508aee8b1e1dac8e69b86f6cee5380664be6
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>