restapi: filtering the api according to ApplicationMode
Currently the Application Mode can contain 3 values.
VirtOnly, GlusterOnly and AllModes.

The value of Application Mode is stored in vdc_options table.
It will be read at the time of login and stored in Current.

In GlusterOnly mode, the links pointing to data center, storage domains
vms will not included. Summary vm details also not included.

In VirtOnly mode, the glusterVolumes under the cluster resource will
be hidden.

Appropriate test cases are written for all 3 modes.

Change-Id: I1cf9bebe85ebc3e321006bc6e45c8ba08bfca6dd
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>