engine: Reduce flood of polling attempts/logs
When SetupNetworksCommand was run and the host didn't lose connectivity
with the engine, we would poll it relentlessly.
This would cause a log flood on both the host and the engine, and also
generates needless traffic over the network.

Added a 0.5 second cooldown period betweeb polling attempts, unless the
polling took more than that time (which probably was caused due to lost
connectivity).
Also reduced the verbosity of the logs on the engine side.

Change-Id: I324e326a6ea8ef92195dd22218bc619a49fe574a
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>