packaging: Don't modifiy limits.conf during setup
Instead of modifying the /etc/security/limits.conf it is better to
drop a small file in the /etc/security/limits.d directory.

For platforms using systemd to manage services the above will not
work ans systemd ignores completely the limits.conf and limits.d
directory. Instead of that the limits have to be specified in the
file defining the service.

This also fixes the name of the user, which should be ovirt
instead of jbossas.

Change-Id: Ib64ce87bd7ddf2749d2d1de85668250d5c59f779
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>