webadmin: replace sysprep tab with intial run
introducing "Initial Run" tab instead of sysprep tab. The tab has
2 sections currently, "General" which is not OS dependent setting. like
timezone and a "Windows" section which has domain dropdown.

A 3rd section is planned for "Linux", when linux sysprep-like functionality
will be available.

The major change here is the ability to set a timezone for Linux OS (OS type
chosen in General Tab) and it will persisted when templating and instantiating
VMs from now on.

timezone is used to set the HW clock of the VM by extracting the utc offset
and passing to VDSM.

Change-Id: Ia5c3536ee83fe456685956f596ac47dd44e3301f
Signed-off-by: Roy Golan <rgolan@redhat.com>