core: Re-enable AddDiskToVmCommandTest
This test was marked with @Ignore a while ago. This patch enables the
test again, and removes the annotation.

In order to achieve this, some subtle refactoring was performed on
AddDiskCommand, AbstractDiskVmCommand and CommandBase (extracting
methods to hide external resources), and PowerMockito was removed from
the test.
A single test which tests the validation of the disk type passed to the
command was removed, since this functionallity is no longer supported.

Change-Id: I3b061a6f381f4e19d6666b24140b0583a244a63d
Signed-off-by: Allon Mureinik <amureini@redhat.com>