core:add support for engine case-insensitive search
Bug origin is the caching mechanism we are using to store queries.
Currently the cash is a Map with a string concatenation of the query text and max results as the key and the SQL generated as the value.
The key should be enhanced to include the Case Sensitive flag value such that 2 identical queries with different values for that flag will be stored in separate cash entries

Change-Id: I129a971e568ea5773a26ef32283ecb16b9252390
Signed-off-by: Eli Mesika <emesika@redhat.com>