webadmin: No red warning when export template that already exists.
Template -> select one -> Export
If this template has already been exported, the popup shuld contain
a red message which warns that this has already been exported.

The problem was in TemplateListModel.showWarningOnExistingTemplates
method which used synchronous queries to retrive storage_pool
and templates on the export domain which does not work on the
GWT envirotnment.

Fixed by extracting and generalizing the correct logic from VmListModel
to a common parent VmBaseListModel and calling this common logic.

Added also a progress animation, until all the required data are
fetched from server, so will not happen, that the window will not
contain the warning message for a while and will look like there
is no warning.

Change-Id: Id96054d4c464ae0373cfa3c2487d8021e2971568
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>