core: UnPowerMock and clean pools with vm commands
Removed PowerMock depedency and replaced it with directly spying the
relevant methods in the tests of commands relevant to manipulating pools
which contain VMs.

Some refactoring in the commands themselves and their base classes was
required in order to achieve this functionality - mainly extracting DAOs
and other methods which access external resources to their own helper
methods.

Change-Id: I43cdf43682d3ab2a3e75ca965490982888b93763
Signed-off-by: Allon Mureinik <amureini@redhat.com>