core: Network is created as bridge-less (#833028)
https://bugzilla.redhat.com/833028

When network is being created over bond, no value is being set for the
vm network attribue of the AddNetwork command which causes the network
to be created as bridge-less.

The patch uses a c'tor of the AddNetwork command which is capable to get
the network entity and set its attributes properly.

Change-Id: Id52d9d618603cd64c5d19650f6437d9d92c9a75e
Signed-off-by: Moti Asayag <masayag@redhat.com>