webadmin: Fix getFriendlyVersion
getFriendlyVersion should return 3.1 for vdsm version
4.10.
Most of the usages of getFriendlyVersion compare the result
with the minimum compatibility version (check that it is equal
or greater than the mimimum compatbility version for a feature)
Prior to this patch, 4.10 would have resulted as version 4.10
so the comparsion would have returned a value greater than 0,
now the Version returned from getFriendlyVersion will be 3.1
(as for 4.9 it returned 3.0).
So the comparsion would also return a value greater than 0
(So no regression here).
The only change is that at UI for host you will see vdsm version
3.1.x.y instead of 4.10.x.y
This is an interim patch until we start working with RpmVersion
and then we should revisit this.

Change-Id: I88e55c73d4b7747d8af8dcda5e112694aeea3030
Acked-By:
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>