wedadmin: switching engine net to new nic causes error
witching rhevm network to different NIC throws
exception NETWORK_ATTACH_ILLEGAL_GATEWAY
The engine throws this error because the old nic keeps
the old default gateway attribute value even there
is no network attached to it. The engine checks if the
nic with the default gateway is management- it isn't
so it throws the error.

Change-Id: I05752b8e5283f86778155e9fa04d3e52215cf427
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>