core:Validate snapshot when detach a disk
When detaching a disk, there should be a validation if the disk is part
of a snapshot or not.
That is because the disk can be detached and attached to other VM and
if the original VM will be previewed, we will have two similar disks in
the system.

I added a validation before detaching a disk whether it contains a
snapshot or not.

Change-Id: I8bef100bc0ccb88dc8c0cdcea9b25084ac7dbfc0
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>