common: Introducing a method to get a direct step.
A direct step is a step that does not have a parent step -
It was created directly under a job.
Example of usage:
We may have a job that has a step for an internal command , which
is a child step of its EXECUTING step.
Since the step for the internal command invokes a command, a child step
of EXECUTING step will be created to represent execution of the internal command.
Since a job holds all its job tree, including for sub step, this method ensure that when needed
the EXECUTING step that was created directly under the job will be returned (and not the EXECUTING step
that was created for the internal command, as explained in the given example)

Change-Id: If002b0e5f3036a17921d907811ff6308b18b1330
Acked-by:
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>