core: move/copy disk entity id fix
when moving image the entity id of the command's parameters, should
be the dest disk uuid,
because when copying the same image, more than once,
only one task is getting cleared (all of the copies has the same entity id - the source image).
that caused the vms (minus one) to stay 'locked'.

Change-Id: I286a474ac033f980c8d57618f1637032fa22297a
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>