userportal,webadmin: Querying for events takes too much time
For events sub tab it took too long time to read the data.

The problem was, that the EventListModel.Search() is invoked
when the Events tab is shown. But this method only set up
the query and started the timer, and the events where actually
queried when the timer decided to.

This fix enriched the EventListModel.Search() by explicitly
invoking the refresh command when the events tab is shown.
Note, that this is invoked only when the tab is shown, but
after this, the timer takes care of refreshing the data.

This patch also contains the following fix:
When there are no events for some entity, the progress
animation was on the events sub tab forever.
This has been fixed by sending the LastEvent event
also when no data came back from the server.

Change-Id: Id911179818a36e221ac9157ea929d888b88cc61e
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>