core: Option with empty name and value ...
Option with empty name and value in the request's xml to rest-api should
not be allowed.

The problem is not with the PM options, this is working without a
problem.
There is a leading blank in the UUID of the DC in the request that
caused the exception.
Adding a check for DC ID validity in canDoAction of
ConnectStorageToVdsCommand command.

Change-Id: I8d53cc13dc6669e02020cf1a5bfacb4fc1285b74
Signed-off-by: Eli Mesika <emesika@redhat.com>