core: On scale environment restart jbossas service causes initialization of Engine.ear to fail on timeout
The following problem occurred during restart of jboss, the most long process which take around 50 seconds was
init of MacPoolManger, the reason is a number of vms which were defined at system.
A process of init is retrieves all vms from db and after that tries to retrieve all theirs network interfaces, that
process can take a while.
Proposed solution:
1. Remove unneeded inits from InitBackendServicesOnStartupBean
2. Start init of MacPoolManager at separate thread
3. Optimization of MacPoolManger : start using HashSet instead of List, paying memory, reducing a time
4. Fixing bug with possible ConcurrentModificationException
5. Small clean up

Change-Id: If3bab39681d92a455aef8d642dcb74dde144b396
Signed-off-by: Michael Kublin <mkublin@redhat.com>