engine: filter-out already attached disks (#833076)
https://bugzilla.redhat.com/833076

currently, when attaching disk the gui also shows the already attached disks.
to solve that I've added a new parameter to the SP, vm guid, that filter out
all the disks that are already attached to the vm. in case the new parameter
is null, the same flow is kept (get all shared disks in DC).

Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Change-Id: Ifa416fd6cd8e171295458b79322bb654636413f8