core:upgrade from rhevm is not possible
upgrade from rhevm is not possible (ISO list is empty)
Removing a redundant check in the UI that although query validations
checks in the core passed, did another check and removed items from the
list of ISO returned from core.

The problem was that when the vdsm getCapabilities fail, the OS version
is null/empty. We should enable the user upgrade using available ISO
files event if getCapabilities fails.

Apart from that, the deleted UI code does a check that is already
handled by BLL code and anyway should not be done by UI.

Change-Id: I4992c74b0f05cd3da7829c5749c1b95eba353f8a
Signed-off-by: Eli Mesika <emesika@redhat.com>