core: allow snapshots for imported templates
when importing template, the AllowSnapshot is not initialized,
so by default its false, making all the disks not allowed to have snapshots.
this value is copied to the vms and is problematic.

when importing template, it is safe to assume that the AllowSnapshot is true for the disks,
as we only export disks that AllowSnapshot is true.
(this already happens for vms, this patch fix this for templates)

Change-Id: I12413c8b63a0085bbdb68419318eea77a95c85dd
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>