webadmin: Use an absolute URL for spice CABs
We are currently using URLs relative to the webadmin and user
portal applications in order to locate the spice .cab files. This
forces the spice package to be installed or symlinked inside the
user portal application, which is not nice.

This change uses an URL relative to the root of the server and
configures the file servlet to get the .cab files from their
original location in the file system. It currently assumes that
.cab files are located in /usr/share/spice, but this can be
quickly changed in the web.xml file of the root.war application.

Change-Id: Id22b30f8120e281187572b164b88fa9c260f61dd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>