core: Cleanup and UnPowerMock RunVmCommand[Test]
Extracted the logic of CanRunVM to an external class (VmRunHandler) in
order to decouple RunVmCommand from other commands.
Additional static methods were extracted in VmHadnler and ImagesHandler
to better decouple the command.

Once this was done, the PowerMockito could be removed from
RunVmCommandTest, and the test could generally be cleaned up for better
readability, extendability and performance.

Change-Id: I2650d70f75fddc4fae0c570273d64c0a1ae9d21a
Signed-off-by: Allon Mureinik <amureini@redhat.com>