core: Fixing addSubStep at CommandBase
A subStep should added also in case there is no job associated with the command.
This can happen when a child command is created, and the job from the parent command
was not propagated to it.
This happens for example during AttachUserToVmFromPoolAndRunCommand -
When creating an execution context for child command, it is required to set the step.
However, the job identifies only the parent command that was invoked by the client.
Adding a sub step on steps that are created by the parent command will create NPE
if there will be an attempt to get the parent step from the job. This patch fixes
this issue.

Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>

Change-Id: I0def5f5ffb896ee946ff3472af92c86074977ed1
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>