core: Redirect to index when page not found
Currently when a page in the root web application is not found we
send a 404 error and the application server generates the default
errot page, which is not very nice. This patch changes the root
web application so that it will redirect to the index page.

Note that this will not work for page not found errors in other
applications like webadmin or user portal.

Change-Id: Id6f97e39fea06de211e04aa18371f3b117aebd07
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>