core: Remove disk from snapshots (#828192)
https://bugzilla.redhat.com/828192

When a disk is removed all its snapshots are removed.
However, they are not removed from the snapshot's configuration -
instead, they are marked as ILLEGAL (upon reading!).

Unfortunately, many flows disregard illegal snapshot disks (e.g., BZ
828192), which leads to a bunch of broken flows.

This patch proposes a different approach - since the disk is removed
entirely, with its snapshots, it should also be removed from the
snapshots' configuration, as if it never existed.

Change-Id: I97dedb9b8e8cd4dd0cae6c7b01d4b5678ba0a2d0
Signed-off-by: Allon Mureinik <amureini@redhat.com>