setup: test effective user id instead of 'root' user name
Root user or every other alias will have UID=0, better
test for UID and not user name.

In bash we have $UID, $EUID.

In POSIX sh we have $(id -u).

In Python, there is no need to convert uid to name in order
to test if we run under root.

Change-Id: Iad84de5277439b2a1aac7b5985954c1ad8ffca80
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>