restapi: UnPowerMock REST-API Types tests
Removed PowerMock from REST API Types tests in order to improve the
build performance. This improvement cuts down the build time by 50%-60%.

Since PowerMock was only used to mock the Config object, it could easily
be removed and replaced with the more modern MockConfigRule class.
A side effect of not using the PowerMockRunner anymore is that the
default JUnit runner is used, with forced changing all the tests'
constructors to public instead of protected.

Change-Id: I44d281c8f8c8568967ef657409d597f6b07bffe0
Signed-off-by: Allon Mureinik <amureini@redhat.com>