core: db changes to add association between tasks and entities
This patch is the beginning of entity association to
tasks that relate to it.
The patch introduces the table asyc_task_entities that holds in
each record ID of async task, ID of entity ID and the type of the entity.
It adds a psql function, used by AsyncTaskDao, to get all taskIDs for a
given entity.

Change-Id: I7a711e932ffb7bcc6920a3456ac97f8b8fc18a76
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>