core: Fixing canDoAction of DeactiveStorageDomainCommand
With this fix - tasks are associated with their entities, there is no need to query
VDSM for HSM task statuses.
This both saves network communication, and in addition, the VDSM call does not provide
any information on the storage domain, so using it allows only to check if the data
center has running tasks.
With the fix, a query to db to check if the storage domain is associated with
tasks replaces the described above VDSM call.

Change-Id: Ic5841d558a7a5267a67cc9547bac187f9773b87b
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>