core: Refactoring AsyncTaskManager
This patch performs the following refactors:
a. Refactors createTask to return an SPMAsyncTask object (SPMAsyncTask is the base class for tasks)
and not the task Id (guid).
b. Takes out DB related code from the CTOR of SPMAsyncTask.
c. Introduces AsyncTaskUtils to hold some common logic for async tasks
d. Introduces entity IDs and entity type to SPMAsyncTask in order to store entity association with tasks

Change-Id: Ia820fd378331b1360cb2766feda80d8c2878310e
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>