packaging, tools: Generate engine-config.xml from template
This change converts the engine-config.xml file into a
ovirt-engine.xml.in template that is copied from
/usr/share/ovirt-engine/service to /var/cache/ovirt-engine when
the engine starts.

The template is processed and the variables are replaced by the
values in /etc/sysconfig/ovirt-engine.

This change is combined with changes to engine-setup, the tools
(engine-config and engine-manage-domains) and the notifier so that
they don't manipulate the engine-config.xml file, but the simpler
/etc/sysconfig/ovirt-engine file.

Change-Id: I36c8ff572307a31e5f9f9749e6c3cc41858a2765
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>