core:vm migration in case of failure ...(#818289)
vm migration in case of failure is not logged in task manager

https://bugzilla.redhat.com/show_bug.cgi?id=818289

When we have Host A and Host B where Host A is SPM and we have HA VM
running on Host A, if Host A is going to non-operational (can not access
the storage for example) the HA VM will be migrated to Host B and Host B
will become the SPM but we will not see the Migration under the
SetNonOperational job.
The reason is that the SetNonOperational ExceutionContext is not passed
to MigrateAllVms method in order to use this context as the parent job
context and to associate all migration steps with the SetNonOperational
job.
This patch changes the MigrateAllVms call to pass the parent
SetNonOperational job ExceutionContext, this context is used as the
parent ExceutionContext when migration steps are created for each VM.

Change-Id: I437893ccf373af863810696d14ce8fd1ffb0253b
Signed-off-by: Eli Mesika <emesika@redhat.com>