core: Validate the connectivity timeout for SetupNetworks
The patch validate the value of the connectivity timeout when it is
passed by the clients for SetupNetworks command.

The legal range for the timeout is between 1 and 120 seconds. If the
check connectivity property is set to true, but no timeout is passed,
the command will be executed with the default timeout.

Change-Id: I7e03ec8dc60710780df9bf7b711d639b8c80327e
Signed-off-by: Moti Asayag <masayag@redhat.com>