webadmin: Use AspectJ annotations instead of compiler
The advantage of using the AspectJ annotations style instead of
the AspectJ language and compiler is that there is no need to use
an special maven plugin to do the build of the aspects.

This also means that there is no need to use any special Eclipse
plugin to support that maven pluging.

Change-Id: I19fceace94978ada932f214f936041e1257fafa9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>