core: MAC address released when MAC is used (#839201)
https://bugzilla.redhat.com/839201

The patch prevents from releasing an allocated MAC address before the
completion of the can-do-action. The MAC address will be released as the
last step of the command execution. However if from some reason the
command will fail to complete its transaction, the rollback will try to
allocate the MAC address again and free the allocated one.

Change-Id: Ifb5109831cf1e290628e579add0ed060393b0d99
Signed-off-by: Moti Asayag <masayag@redhat.com>