packaging: Run with OpenJDK 7 regardless of what is installed
This changes the start/stop script so that it will use OpenJDK 7
regardless of what other versions of the Java virtual machine are
installed.

If the Java virtual machine happens to be installed in a different
place that the one used by default by the script it can be
specified in the /etc/sysconfig/ovirt-engine  with the JAVA_HOME
variable.

Change-Id: I44b5b4e20a134d1624acf1258c59b73d8d2382c6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>