database: do not use git when building
If we are not building on development machine, for example we
use rpmbuild -tb <tarball>, we probably do not have git or
we do not have git checkout to query.

There is no actual need to use git, as we can use relative
directories.

Cleanup the check_for_duplicate_upgrade_scripts.sh script.

Change-Id: If06ea52efb4da7c51dad8fc2b32043082ce13263
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>