core: Add configured range validator
The patch introduces configured range validator which could be set on
business entities properties to enforce the valid range of a given
input.

The declaration of the validator requires a configuration value to
specify the upper limit of the range. The lower limit could be specified
either by a configuration value or by setting a constant or using the
default (zero).

Change-Id: I23642aa79b263019acbdde4ee44e27f5a0ac992e
Signed-off-by: Moti Asayag <masayag@redhat.com>